Setup Webpack devServer and made some corrections to seeding script
