+++ .babelrc
@@ -1,6 +1,7 @@
 ﻿{
   "presets": [
-    ["@babel/preset-env", "@babel/preset-react"]
+    "@babel/preset-env",
+    "@babel/preset-react"
   ],
   "plugins": ["react-hot-loader/babel"]
 }
