+++ Components/App.js
@@ -1,38 +1,49 @@
-﻿import React, {Component} from "react";
-import { hot } from 'react-hot-loader/root';
-const axios = require('axios');
+﻿// import React, {Component} from "react";
+// import { hot } from 'react-hot-loader/root';
+// const axios = require('axios');
+import _ from 'lodash';
 
-class App extends Component{
-  constructor(props) {
-    super(props);
-    this.state = {name: 'Unknown'};
+function component() {
+  let element = document.createElement('div');
 
-    // this.handleClick = this.handleClick.bind(this);
-  }
+  element.innerHTML = _.join(['Hello', 'webpack'], ' ');
 
-  componentDidMount() {
-    axios.get('/listings?id=thfUjkN4Tv')
-      .then(function (response) {
-        this.setState({
-          name: res.name
-        })
-        console.log(response);
-      })
-      .catch(function (error) {
-        // handle error
-        console.log(error);
-      })
-      // .then(function () {
-      // });
-  }
-
-  render(){
-    return(
-      <div className="App">
-        <h1> Hello, World! </h1>
-      </div>
-    );
-  }
+  return element;
 }
 
-export default hot(App)
+document.body.appendChild(component());
+
+// class App extends Component{
+//   constructor(props) {
+//     super(props);
+//     this.state = {name: 'Unknown'};
+
+//     // this.handleClick = this.handleClick.bind(this);
+//   }
+
+//   componentDidMount() {
+//     axios.get('/listings?id=thfUjkN4Tv')
+//       .then(function (response) {
+//         this.setState({
+//           name: res.name
+//         })
+//         console.log(response);
+//       })
+//       .catch(function (error) {
+//         // handle error
+//         console.log(error);
+//       })
+//       // .then(function () {
+//       // });
+//   }
+
+//   render(){
+//     return(
+//       <div className="App">
+//         <h1> Hello, World! </h1>
+//       </div>
+//     );
+//   }
+// }
+
+// export default hot(App)
