--- index.html
@@ -1,14 +1,0 @@
-﻿<!doctype html>
-  <html>
-   <head>
-    <title>Getting Started</title>
-    <script src="https://unpkg.com/lodash@4.16.6"></script>
-   </head>
-   <body>
-      <div id="root"></div>
-    <script src="./src/index.js"></script>
-    <script src="./dist/main.js"></script>
-    <!-- <script src="http://localhost:8080/dist/js/vendors.js"></script>
-    <script src="http://localhost:8080/dist/js/app.bundle.js"></script> -->
-   </body>
-  </html>
