+++ package.json
@@ -5,18 +5,18 @@
   "author": "Michael Ryan",
   "license": "MIT",
   "engines": {
-    "node": ">=6.13.0",
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/br0kentable/Restaurant_details"
-    }
+    "node": ">=6.13.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/br0kentable/Restaurant_details"
   },
   "scripts": {
     "start": "webpack-dev-server --hot --mode development --open",
     "dev": "webpack --mode development",
     "build": "webpack",
     "prod": "webpack --mode production",
-    "seed": "node seeding.js"
+    "seed": "node Back/Seeding.js"
   },
   "dependencies": {
     "axios": "^0.18.0",
