--- src/index.js
@@ -1,12 +1,0 @@
-﻿import _ from 'lodash';
-
-function component() {
-  let element = document.createElement('div');
-
-  // Lodash, currently included via a script, is required for this line to work
-  element.innerHTML = _.join(['Hello', 'webpack'], ' ');
-
-  return element;
-}
-
-document.body.appendChild(component());
