+++ webpack.config.js
@@ -1,15 +1,20 @@
 ﻿const path = require('path');
+const webpack = require('webpack');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: './Components/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist')
+  },
   mode: "development",
   module: {
     rules: [
-      {
-        test: /\.jsx?$/,
-        loaders: ["react-hot"],
-        include: path.join(__dirname, "public")
-      },
+      // {
+      //   test: /\.jsx?$/,
+      //   loaders: ["react-hot"],
+      //   include: path.join(__dirname, "public")
+      // },
       {
         test: /\.(js|jsx)$/,
         exclude: /(node_modules)/,
@@ -21,9 +26,12 @@ module.exports = {
       }
     ]
   },
-  resolve: { extensions: ["*", ".js", ".jsx"] },output: {
-    filename: 'main.js',
-    path: path.resolve(__dirname, 'dist')
+  resolve: { extensions: ["*", ".js", ".jsx"] },
+  devtool: 'inline-source-map',
+  devServer: {
+    contentBase: './dist',
+    publicPath: './dist',
+    hot: true
   },
   plugins: [
     new webpack.HotModuleReplacementPlugin()
